Changed all component methods
Changed every method under the Component trait to take a handle to themselves. Added code to execute postupdate to the main loop. Made handle_mutex take a reference instead of an owned value. Changed the way objects and components marked for death are handled by passing the proper handle instead of just the code.